length type has been changed as demand

+++ doubly_linked_lists/0-print_dlistint.c
@@ -1,10 +1,16 @@
 #include "lists.h"
-
+/**
+ * print_dlistint - it prints our double linked list
+ * @h: it is our head
+ * Return: it returns lentgth of our double linked list
+ *
+ *
+ */
 size_t print_dlistint(const dlistint_t *h)
 {
-int length = 0;
+size_t length = 0;
 
-while(h)
+while (h)
 {
 	printf("%d", h->n);
 	h = h->next;
